# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing data on oncological markers.",
        "title": "OncomarkerAPI",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/blood-count-value": {
            "get": {
                "description": "Retrieves a list of all blood count values.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BloodCountValue"
                ],
                "summary": "Get blood count value list",
                "responses": {
                    "200": {
                        "description": "Blood count value list",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/model.BloodCountValue"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates an existing blood count value entry.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BloodCountValue"
                ],
                "summary": "Update blood count value",
                "parameters": [
                    {
                        "description": "Blood count value data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.BloodCountValue"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated blood count value data",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new blood count value entry.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BloodCountValue"
                ],
                "summary": "Create blood count value",
                "parameters": [
                    {
                        "description": "Blood count value data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.BloodCountValue"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Created blood count value data",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/blood-count-value/blood-count/{blood_count_id}": {
            "get": {
                "description": "Retrieves a list of blood count values associated with a specific blood count.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BloodCountValue"
                ],
                "summary": "Get blood count value list by blood count",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Blood count ID",
                        "name": "blood_count_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Blood count value list",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/model.BloodCountValue"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/blood-count-value/disease/{disease_id}": {
            "get": {
                "description": "Retrieves a list of blood count values associated with a specific disease.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BloodCountValue"
                ],
                "summary": "Get blood count value list by disease",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Disease ID",
                        "name": "disease_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Blood count value list",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/model.BloodCountValue"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/blood-count-value/{disease_id}/{blood_count_id}": {
            "get": {
                "description": "Retrieves a blood count value by its ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BloodCountValue"
                ],
                "summary": "Get blood count value by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Disease ID",
                        "name": "disease_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Blood count ID",
                        "name": "blood_count_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Blood count value",
                        "schema": {
                            "$ref": "#/definitions/model.BloodCountValue"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes an existing blood count value entry.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BloodCountValue"
                ],
                "summary": "Delete blood count value",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Disease ID",
                        "name": "disease_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Blood count ID",
                        "name": "blood_count_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted blood count value data",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "model.BloodCountValue": {
            "type": "object",
            "required": [
                "blood_count",
                "coefficient",
                "disease"
            ],
            "properties": {
                "blood_count": {
                    "description": "Type of blood count.",
                    "type": "string"
                },
                "coefficient": {
                    "description": "Coefficient value.",
                    "type": "number"
                },
                "description": {
                    "description": "Description of the blood count value. Optional.",
                    "type": "string"
                },
                "disease": {
                    "description": "Name of the disease.",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}